Adding directX Tex library

+++ Direct3D11Renderer/src/Core/Window.cpp
@@ -47,7 +47,7 @@ HINSTANCE Window::WindowClass::GetInstance() noexcept
 }
 
 // Window
-Window::Window(int width, int height, const WCHAR* name)
+Window::Window(float width, float height, const WCHAR* name)
 	: width(width), height(height), hwnd(nullptr)
 {
 	RECT wr;
